Renaming files
Renamed several files to be clearer on their intent and to structure the
navigation in a clearer form.

* Org-mode
Changed from Literate programming to Org-mode to take a wider view of Org-mode,
rather than just the literate programming approach.

* iedit
Renamed multiple cursors to find and replace, as that is the action that is
being described.