Update - enhancing clojure layer experience
Sayid debugger package and clj-refactor are no longer loaded by default.

Sayid has caused a few issues with Cider recently so it is left to the user to decide
if it is useful for them.

clj-refactor has not been updated in several years and some of its functionality
is moving into clojure-mode.

Both these packages can be included by defining their package names as
:variables on the clojure layer