New section - Linting
Quick overview of linting tools I use for my Clojure projects, Joker and Eastwood.

Joker is simpler to use, thanks to the clojure-lint layer.  It does use an
external binary, which I placed on my existing executable path.  Joker uses a
sub-set of Clojure so it may give a few inaccuracies, these are usually false
positives on things like macros.  There is a way to tell joker to ignore certain
symbols though.

Still a bit of work to do on these sections, but they cover the basics.