Re-organised the refactor section - simplifying the overall sections
Moved clj-refactor as a sub-section of the refactor section as part of
restructuring the overall book to have fewer main sections.

Adding a new refactor section introduction page, to cover the main approaches to
refactor code in Spacemacs.

Will add more content in this section, especially around iedit, narrowing and
helm.