Updated Linter and static analysis sections
Simplified the section on linting, recommending just clj-kondo for live linting.

Renamed Eastwood linting section to batch linting, to express that Eastwood is
better used as a check at certain points in the development cycle, eg. before
sharing commits or as part of a CI/CD process.

Moved joker section to the reference section, creating a new Alternative tools
section.

Resolves #66