Vim - update searching tips to use helm-swoop
Using `/` to search leads to strange effects with Vim editing if `RET` is not
pressed to complete the search term.

Swoop seems more effective, as it presents a list of lines that contain the
matching search pattern, so its easy to jump to exactly where you want to go.

In the `/` search you have to navigate one by one and the search term stays
highlighted if you do not clear the search.